refactor:配置文件缓存逻辑重构
