+++ test/suit/test_suit.go
@@ -1001,52 +1001,6 @@ func (d *DiscoverTestSuit) CleanCircuitBreaker(id, version string) {
 
 // 彻底删除熔断规则发布记录
 func (d *DiscoverTestSuit) CleanCircuitBreakerRelation(name, namespace, ruleID, ruleVersion string) {
-
-	if d.Storage.Name() == sqldb.STORENAME {
-		func() {
-			tx, err := d.Storage.StartTx()
-			if err != nil {
-				panic(err)
-			}
-
-			dbTx := tx.GetDelegateTx().(*sqldb.BaseTx)
-
-			defer rollbackDbTx(dbTx)
-
-			str := "delete from circuitbreaker_rule_relation using circuitbreaker_rule_relation, service " +
-				"where service_id = service.id and name = ? and namespace = ? and rule_id = ? and rule_version = ?"
-			if _, err := dbTx.Exec(str, name, namespace, ruleID, ruleVersion); err != nil {
-				panic(err)
-			}
-
-			commitDbTx(dbTx)
-		}()
-	} else if d.Storage.Name() == boltdb.STORENAME {
-		func() {
-			releations, err := d.Storage.GetCircuitBreakerRelation(ruleID, ruleVersion)
-			if err != nil {
-				panic(err)
-			}
-			tx, err := d.Storage.StartTx()
-			if err != nil {
-				panic(err)
-			}
-
-			dbTx := tx.GetDelegateTx().(*bolt.Tx)
-
-			for i := range releations {
-				if err := dbTx.Bucket(
-					[]byte(tblCircuitBreakerRelation)).DeleteBucket([]byte(releations[i].ServiceID)); err != nil {
-					if !errors.Is(err, bolt.ErrBucketNotFound) {
-						rollbackBoltTx(dbTx)
-						panic(err)
-					}
-				}
-			}
-
-			commitBoltTx(dbTx)
-		}()
-	}
 }
 
 func (d *DiscoverTestSuit) ClearTestDataWhenUseRDS() error {
